create viz plot tool
