+++ scripts/scil_dti_metrics.py
@@ -27,7 +27,6 @@ Formerly: scil_compute_dti_metrics.py
 import argparse
 import logging
 
-import matplotlib.pyplot as plt
 import nibabel as nib
 import numpy as np
 
@@ -54,6 +53,7 @@ from scilpy.gradients.bvec_bval_tools import (check_b0_threshold,
                                               is_normalized_bvecs,
                                               normalize_bvecs)
 from scilpy.utils.filenames import add_filename_suffix, split_name_with_nii
+from scilpy.viz.plot import plot_residuals
 
 logger = logging.getLogger("DTI_Metrics")
 logger.setLevel(logging.INFO)
@@ -147,75 +147,6 @@ def _build_arg_parser():
     add_verbose_arg(p)
 
     return p
-
-
-def _plot_residuals(args, data_diff, mask, R_k, q1, q3, iqr, residual_basename):
-    # Showing results in graph
-    # Note that stats will be computed manually and plotted using bxp
-    # but could be computed using stats = cbook.boxplot_stats
-    # or pyplot.boxplot(x)
-    if mask is None:
-        logging.info("Outlier detection will not be performed, since no "
-                     "mask was provided.")
-
-    # Initializing stats as a List[dict]
-    stats = [dict.fromkeys(['label', 'mean', 'iqr', 'cilo', 'cihi',
-                            'whishi', 'whislo', 'fliers', 'q1',
-                            'med', 'q3'], [])
-             for _ in range(data_diff.shape[-1])]
-
-    nb_voxels = np.count_nonzero(mask)
-    percent_outliers = np.zeros(data_diff.shape[-1], dtype=np.float32)
-    for k in range(data_diff.shape[-1]):
-        stats[k]['med'] = (q1[k] + q3[k]) / 2
-        stats[k]['mean'] = R_k[k]
-        stats[k]['q1'] = q1[k]
-        stats[k]['q3'] = q3[k]
-        stats[k]['whislo'] = q1[k] - 1.5 * iqr[k]
-        stats[k]['whishi'] = q3[k] + 1.5 * iqr[k]
-        stats[k]['label'] = k
-
-        # Outliers are observations that fall below Q1 - 1.5(IQR) or
-        # above Q3 + 1.5(IQR) We check if a voxel is an outlier only if
-        # we have a mask, else we are biased.
-        if args.mask is not None:
-            x = data_diff[..., k]
-            outliers = (x < stats[k]['whislo']) | (x > stats[k]['whishi'])
-            percent_outliers[k] = np.sum(outliers) / nb_voxels * 100
-            # What would be our definition of too many outliers?
-            # Maybe mean(all_means)+-3SD?
-            # Or we let people choose based on the figure.
-            # if percent_outliers[k] > ???? :
-            #    logger.warning('   Careful! Diffusion-Weighted Image'
-            #                   ' i=%s has %s %% outlier voxels',
-            #                   k, percent_outliers[k])
-
-    if args.mask is None:
-        fig, axe = plt.subplots(nrows=1, ncols=1, squeeze=False)
-    else:
-        fig, axe = plt.subplots(nrows=1, ncols=2, squeeze=False,
-                                figsize=[10, 4.8])
-        # Default is [6.4, 4.8]. Increasing width to see better.
-
-    medianprops = dict(linestyle='-', linewidth=2.5, color='firebrick')
-    meanprops = dict(linestyle='-', linewidth=2.5, color='green')
-    axe[0, 0].bxp(stats, showmeans=True, meanline=True, showfliers=False,
-                  medianprops=medianprops, meanprops=meanprops)
-    axe[0, 0].set_xlabel('DW image')
-    axe[0, 0].set_ylabel('Residuals per DWI volume. Red is median,\n'
-                         'green is mean. Whiskers are 1.5*interquartile')
-    axe[0, 0].set_title('Residuals')
-    axe[0, 0].set_xticks(range(0, q1.shape[0], 5))
-    axe[0, 0].set_xticklabels(range(0, q1.shape[0], 5))
-
-    if args.mask is not None:
-        axe[0, 1].plot(range(data_diff.shape[-1]), percent_outliers)
-        axe[0, 1].set_xticks(range(0, q1.shape[0], 5))
-        axe[0, 1].set_xticklabels(range(0, q1.shape[0], 5))
-        axe[0, 1].set_xlabel('DW image')
-        axe[0, 1].set_ylabel('Percentage of outlier voxels')
-        axe[0, 1].set_title('Outliers')
-    plt.savefig(residual_basename + '_residuals_stats.png')
 
 
 def main():
@@ -404,6 +335,9 @@ def main():
                      add_filename_suffix(args.pulsation, '_std_b0'))
 
     if args.residual:
+        if mask is None:
+            logging.info("Outlier detection will not be performed, since no "
+                         "mask was provided.")
         # Mean residual image
         S0 = np.mean(data[..., gtab.b0s_mask], axis=-1)
         tenfit2_predict = np.zeros(data.shape, dtype=np.float32)
@@ -435,7 +369,7 @@ def main():
         np.save(add_filename_suffix(res_stats_basename, "_std_residuals"), std)
 
         # Plotting and saving figure
-        _plot_residuals(args, data_diff, mask, R_k, q1, q3, iqr,
+        plot_residuals(data_diff, mask, R_k, q1, q3, iqr,
                         residual_basename)
